Updated tbp% and pDiv% with decimals to resolve #879.
Free Software Free Society

To increase the frequency of commits by ctubio,
you can buy-me-a-drink with a small git-tip at:
  1GitTipB6sY2YSbkH3BJHepUB7Cjymi9R1

To force me to write more source documentation,
you can buy-me-a-drink with a small doc-tip at:
  1KDocsmbhUEh7JgmYjuMH3nDivzSM6dphP

I promise to drink chocolate milk after commit.

To request new features or in case this commit breaks something for you,
please, create a new github issue with all possible information for me,
but never share your API Keys!

+++ src/lib/Krypto.ninja-data.h
@@ -73,11 +73,11 @@ namespace ₿ {
     double            rangePercentage                 = 5.0;
     mFairValueModel   fvModel                         = mFairValueModel::BBO;
     Amount            targetBasePosition              = 1.0;
-    unsigned int      targetBasePositionPercentage    = 50;
+    double            targetBasePositionPercentage    = 50.0;
     Amount            positionDivergence              = 0.9;
     Amount            positionDivergenceMin           = 0.4;
-    unsigned int      positionDivergencePercentage    = 21;
-    unsigned int      positionDivergencePercentageMin = 10;
+    double            positionDivergencePercentage    = 21.0;
+    double            positionDivergencePercentageMin = 10;
     mPDivMode         positionDivergenceMode          = mPDivMode::Manual;
     bool              percentageValues                = false;
     mAutoPositionMode autoPositionMode                = mAutoPositionMode::EWMA_LS;
